refactor(ThemeHeroSection): 비동기 통신 함수 위치 이동
